changed abstractions and responsibilities and renamed change model command base clase
